refactor: decompose the god files for real, and ratchet the split shut
Commit C moved two god files behind barrels. `mcp-adapter.ts` went 2225 → 5
lines and `handlers.ts` arrived at 2070; `daemon.ts` went 3283 → 6 and
`lifecycle.ts` arrived at 3101. The file was renamed, not decomposed, and
nothing in the gate could tell the difference — so the structure it "already
allows" had no reason to ever arrive.

**The ratchet comes first, because it is what makes the rest stick.**
`FILE_SIZE_THRESHOLD` is 800 and `FILE_SIZE_BASELINE` records the 56 files
already over it. The destination is not "every file under 800" — demanding 56
refactors in one change is how a rule gets turned off. It is that the NEXT one
cannot happen: a file absent from the baseline may not exceed the threshold, a
file in it may only shrink, and one that passes under must LEAVE the baseline,
because a stale entry re-authorises the growth the shrink just paid for. It is
declared in `repo-invariants.ts`, so a cone-scoped gate that touched any package
still runs it.

**`handlers.ts` 2070 → six domain modules**, none needing a baseline entry:
handlers 583, project 576, dependencies 356, vitals 315, queue 216, events 128.
The barrel stays, and now earns its place: a façade over seven real modules is
a front door, a façade over one 2070-line file is a redirect.

**`lifecycle.ts` 3101 → 2489**, with its option/registration/handle types in
`daemon/types.ts` and its intervals in `daemon/tunables.ts`. What remains is ONE
function of ~2280 lines — forty inner functions closing over sixty locals. That
is a FUNCTION length problem, and a file-size rule cannot fix it: slicing the
closure into modules that each take a sixty-field context object would satisfy
the number and leave the code worse. It takes a baseline entry that states
exactly that, with a ceiling.

**The guards keyed by path are re-declared, not baselined away.** Two repair
sites moved with the code, so `DECLARED_REPAIR_SITES` names their new homes;
`REPAIR_SCAN_ROOTS` scans `apps/dev/src/mcp` as a directory, because naming the
façade would sweep a file that declares nothing while the refusals it forwards
to went unread. Three tests read `src/daemon.ts` for a call site and found only
re-exports — they read the implementation now, and say why in a line.

490 imports the slicing left behind were removed by measurement rather than by
eye, so `tsconfig-import-hygiene` is back to 2/2 (commit C left it 18 → 20).

apps/dev 436 files / 6076 tests; apps/redskilled 690 passing — the 6 that remain
fail identically on clean origin/main and are not this branch's.
