+++ apps/dev/package.json
@@ -14,7 +14,7 @@
     "bundle:entrypoint": "esbuild ../../packages/shared/entrypoint-cli.ts --bundle --minify --platform=node --format=esm --target=node22 --define:__ENTRYPOINT_ROLE__='\"run:dev\"' --outfile=../../plugins/dev/skills/engineering/afk/bin/afk.mjs --banner:js=\"import { createRequire as __cr } from 'node:module'; const require = __cr(import.meta.url);\"",
     "build": "pnpm run bundle && pnpm run bundle:mcp && pnpm run bundle:red-fetch && pnpm run bundle:entrypoint",
     "test": "NODE_OPTIONS=--max-old-space-size=3072 vitest run",
-    "test:invariants": "NODE_OPTIONS=--max-old-space-size=3072 vitest run tests/toon-json-guard.test.ts tests/extinct-source-guard.test.ts tests/custody-verdict-park-guard.test.ts tests/shipped-primitive-guard.test.ts tests/host-owns-birth-guard.test.ts tests/shipped-binary-guard.test.ts tests/declared-wait-guard.test.ts tests/lane-run-mode-guard.test.ts tests/bare-invocation-guard.test.ts tests/version-train-guard.test.ts tests/repair-guard.test.ts tests/truecolor-extinction-guard.test.ts tests/codex-manifest-generator.test.ts tests/github-read-route-guard.test.ts",
+    "test:invariants": "NODE_OPTIONS=--max-old-space-size=3072 vitest run tests/toon-json-guard.test.ts tests/extinct-source-guard.test.ts tests/custody-verdict-park-guard.test.ts tests/shipped-primitive-guard.test.ts tests/host-owns-birth-guard.test.ts tests/shipped-binary-guard.test.ts tests/declared-wait-guard.test.ts tests/lane-run-mode-guard.test.ts tests/bare-invocation-guard.test.ts tests/version-train-guard.test.ts tests/repair-guard.test.ts tests/truecolor-extinction-guard.test.ts tests/codex-manifest-generator.test.ts tests/github-read-route-guard.test.ts tests/file-size-guard.test.ts",
     "test:canary": "NODE_OPTIONS=--max-old-space-size=3072 vitest run tests/mcp-lane-canary.test.ts tests/mcp-lane-canary-e2e.test.ts"
   },
   "dependencies": {
