+++ apps/dev/src/mcp/operations.ts
@@ -16,7 +16,6 @@ import type {
   HitlResolveInput,
   DailyReviewInput,
   LandBranchInput,
-  QueueStatusInput,
   RequeueToolInput,
   RespondToolInput,
   RetakeToolInput,
