+++ apps/redskilled/tests/statusline-string.test.ts
@@ -333,7 +333,8 @@ describe("one renderer, machine-wide", () => {
     "apps/redskilled/src/statusline-render.ts",
     "apps/redskilled/src/dashboard-render.ts",
     "apps/redskilled/src/statusline-payload.ts",
-    "apps/redskilled/src/daemon.ts",
+    // The implementation, not the façade `daemon.ts`, which forwards only.
+    "apps/redskilled/src/daemon/lifecycle.ts",
     "apps/redskilled/src/protocol.ts",
     "apps/redskilled/src/client.ts",
   ]);
